Split out Cells into its own test-utils crate
