+++ Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 resolver = "2"
 members = [
+  "test-utils/cells",
   "test-utils/strategy-ext",
   "tiledb/api",
   "tiledb/common",
@@ -38,6 +39,8 @@ paste = "1.0"
 proptest = { version = "1.0.0" }
 serde = { version = "1", features = ["derive"] }
 serde_json = { version = "1", features = ["float_roundtrip"] }
+strategy-ext = { path = "test-utils/strategy-ext", version = "0.1.0" }
+tempfile = { version = "3" }
 thiserror = { version = "1" }
 tiledb-api = { path = "tiledb/api", version = "0.1.0" }
 tiledb-common = { path = "tiledb/common", version = "0.1.0" }
