+++ tiledb/serde/src/array/dimension/strategy.rs
@@ -5,12 +5,12 @@ use num_traits::{Bounded, FromPrimitive, Num};
 use proptest::prelude::*;
 use proptest::strategy::ValueTree;
 
+use strategy_ext::StrategyExt;
 use tiledb_common::array::ArrayType;
 use tiledb_common::datatype::physical::BitsOrd;
 use tiledb_common::datatype::strategy::*;
 use tiledb_common::datatype::Datatype;
 use tiledb_common::physical_type_go;
-use tiledb_test_utils::strategy::StrategyExt;
 use tiledb_utils::numbers::{
     NextDirection, NextNumericValue, SmallestPositiveValue,
 };
