+++ tiledb/test-utils/Cargo.toml
@@ -5,11 +5,12 @@ rust-version.workspace = true
 version.workspace = true
 
 [dependencies]
-anyhow = { workspace = true }
+anyhow = { workspace = true, optional = true }
 nix = { version = "0", features = ["signal"], optional = true }
-proptest = { workspace = true }
-tempfile = "3"
+paste = { workspace = true, optional = true }
+tempfile = { workspace = true, optional = true }
 
 [features]
-default = ["signal"]
+default = ["uri"]
 signal = ["dep:nix"]
+uri = ["dep:anyhow", "dep:tempfile"]
