+++ tiledb/test-utils/src/lib.rs
@@ -1,11 +1,11 @@
-extern crate proptest;
-
 #[cfg(feature = "signal")]
 pub mod signal;
 
+#[cfg(feature = "uri")]
 pub mod uri_generators;
 
 #[cfg(feature = "signal")]
 pub use signal::*;
 
+#[cfg(feature = "uri")]
 pub use uri_generators::{get_uri_generator, TestArrayUri};
